changes from sanders pr review
