+++ api/hubd/client/client.go
@@ -2,7 +2,9 @@ package client
 
 import (
 	"context"
+	"fmt"
 
+	filrewardspb "github.com/textileio/textile/v2/api/filrewardsd/pb"
 	pb "github.com/textileio/textile/v2/api/hubd/pb"
 	"google.golang.org/grpc"
 )
@@ -166,4 +168,47 @@ func (c *Client) IsOrgNameAvailable(ctx context.Context, name string) (*pb.IsOrg
 func (c *Client) DestroyAccount(ctx context.Context) error {
 	_, err := c.c.DestroyAccount(ctx, &pb.DestroyAccountRequest{})
 	return err
+}
+
+func (c *Client) ListFilRewards(ctx context.Context, opts ...ListFilRewardsOption) ([]*filrewardspb.Reward, bool, int64, error) {
+	req := &pb.ListFilRewardsRequest{}
+	for _, opt := range opts {
+		opt(req)
+	}
+	res, err := c.c.ListFilRewards(ctx, req)
+	if err != nil {
+		return nil, false, 0, fmt.Errorf("calling list fil rewards rpc: %v", err)
+	}
+	return res.Rewards, res.More, res.MoreToken, nil
+}
+
+func (c *Client) ClaimFil(ctx context.Context, amount int64) error {
+	req := &pb.ClaimFilRequest{
+		Amount: amount,
+	}
+	_, err := c.c.ClaimFil(ctx, req)
+	if err != nil {
+		return fmt.Errorf("calling claim fil rpc: %v", err)
+	}
+	return nil
+}
+
+func (c *Client) ListFilClaims(ctx context.Context, opts ...ListFilClaimsOption) ([]*filrewardspb.Claim, bool, int64, error) {
+	req := &pb.ListFilClaimsRequest{}
+	for _, opt := range opts {
+		opt(req)
+	}
+	res, err := c.c.ListFilClaims(ctx, req)
+	if err != nil {
+		return nil, false, 0, fmt.Errorf("calling list fil claims rpc: %v", err)
+	}
+	return res.Claims, res.More, res.MoreToken, nil
+}
+
+func (c *Client) FilRewardsBalance(ctx context.Context) (*pb.FilRewardsBalanceResponse, error) {
+	res, err := c.c.FilRewardsBalance(ctx, &pb.FilRewardsBalanceRequest{})
+	if err != nil {
+		return nil, fmt.Errorf("calling fil rewards balance rpc: %v", err)
+	}
+	return res, nil
 }
