+++ api/usersd/pb/usersd.proto
@@ -3,7 +3,6 @@ package api.usersd.pb;
 option go_package = "github.com/textileio/textile/v2/api/usersd/pb";
 
 import "api/billingd/pb/billingd.proto";
-import "api/filrewardsd/pb/filrewards.proto";
 
 message ListThreadsRequest {}
 
@@ -170,51 +169,6 @@ message ArchiveRetrievalLogsRequest {
 message ArchiveRetrievalLogsResponse {
     string msg = 1;
 }
-  
-message ListFilRewardsRequest {
-    bool unlocked_by_dev = 1;
-    api.filrewardsd.pb.RewardType reward_type_filter = 2;
-    bool ascending = 3;
-    int64 more_token = 4;
-    int64 limit = 5;
-}
-
-message ListFilRewardsResponse {
-    repeated api.filrewardsd.pb.Reward rewards = 1;
-    bool more = 2;
-    int64 more_token = 3;
-}
-
-message ClaimFilRequest {
-    int64 amount = 1;
-}
-
-message ClaimFilResponse {
-}
-
-message ListFilClaimsRequest {
-    bool claimed_by_dev = 1;
-    api.filrewardsd.pb.ClaimState state_filter = 2;
-    bool ascending = 3;
-    int64 more_token = 4;
-    int64 limit = 5;
-}
-
-message ListFilClaimsResponse {
-    repeated api.filrewardsd.pb.Claim claims = 1;
-    bool more = 2;
-    int64 more_token = 3;
-}
-
-message FilRewardsBalanceRequest {
-}
-
-message FilRewardsBalanceResponse {
-    int64 rewarded = 1;
-    int64 pending = 2;
-    int64 claimed = 3;
-    int64 available = 4;
-}
 
 service APIService {
     rpc GetThread(GetThreadRequest) returns (GetThreadResponse) {}
@@ -237,10 +191,4 @@ service APIService {
     // Archive Retrieval
     rpc ArchiveRetrievalLs(ArchiveRetrievalLsRequest) returns (ArchiveRetrievalLsResponse){}
     rpc ArchiveRetrievalLogs(ArchiveRetrievalLogsRequest) returns (stream ArchiveRetrievalLogsResponse) {}
-
-    // Fil Rewards
-    rpc ListFilRewards(ListFilRewardsRequest) returns (ListFilRewardsResponse) {}
-    rpc ClaimFil(ClaimFilRequest) returns (ClaimFilResponse) {}
-    rpc ListFilClaims(ListFilClaimsRequest) returns (ListFilClaimsResponse) {}
-    rpc FilRewardsBalance(FilRewardsBalanceRequest) returns (FilRewardsBalanceResponse) {}
 }
