+++ core/core.go
@@ -400,14 +400,14 @@ func NewTextile(ctx context.Context, conf Config, opts ...Option) (*Textile, err
 			BillingClient:       t.bc,
 			PowergateClient:     t.pc,
 			PowergateAdminToken: conf.PowergateAdminToken,
+			FilRewardsClient:    t.frc,
 		}
 		us = &usersd.Service{
-			Collections:      t.collections,
-			Mail:             t.mail,
-			BillingClient:    t.bc,
-			FilRetrieval:     t.filRetrieval,
-			PowergateClient:  t.pc,
-			FilRewardsClient: t.frc,
+			Collections:     t.collections,
+			Mail:            t.mail,
+			BillingClient:   t.bc,
+			FilRetrieval:    t.filRetrieval,
+			PowergateClient: t.pc,
 		}
 	}
